requirements.txt
pillow
pillow-avif-plugin
tkinterdnd2
